backend/requirements.txt
# Core Flask framework
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SocketIO>=5.3.6

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
imutils>=0.5.4

# AI/ML Libraries  
ultralytics>=8.0.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Communication
python-socketio>=5.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Utilities
matplotlib>=3.7.0
eventlet>=0.33.3

# Note: Some packages removed for Python 3.13 compatibility
# face-recognition and dlib have compilation issues on newer Python versions
# We'll use OpenCV's built-in face detection instead